requirements.txt
streamlit
streamlit-drawable-canvas
openpyxl
Pillow
firebase-admin
